requirements.txt
amqp==2.5.2
asgiref==3.2.3
attrs==19.3.0
billiard==3.6.3.0
bootstrap4==0.1.0
celery==4.4.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.9
django-crispy-forms==1.9.1
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.6.0
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
kombu==4.6.8
Markdown==3.2.2
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.4
Pillow==7.0.0
psycopg2==2.8.4
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
pyTelegramBotAPI==3.6.7
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
swagger-spec-validator==2.6.0
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
whitenoise==5.0.1
zipp==3.1.0
